STDC-re/requirements.txt
easydict
matplotlib
numpy
onnx
opencv-python
Pillow
protobuf
scipy
thop
torch
torchvision
tqdm

